requirements.txt
python-decouple~=3.8